refactor: share invoice subtotal/tax/total calculation
Extract the duplicated line-item total logic from create and update into
calculateTotals so both paths use one implementation.
