+++ backend/src/controllers/appControllers/invoiceController/create.js
@@ -5,6 +5,7 @@ const Model = mongoose.model('Invoice');
 const { calculate } = require('@/helpers');
 const { increaseBySettingKey } = require('@/middlewares/settings');
 const schema = require('./schemaValidate');
+const calculateTotals = require('./calculateTotals');
 
 const create = async (req, res) => {
   let body = req.body;
@@ -21,21 +22,7 @@ const create = async (req, res) => {
 
   const { items = [], taxRate = 0, discount = 0 } = value;
 
-  // default
-  let subTotal = 0;
-  let taxTotal = 0;
-  let total = 0;
-
-  //Calculate the items array with subTotal, total, taxTotal
-  items.map((item) => {
-    let total = calculate.multiply(item['quantity'], item['price']);
-    //sub total
-    subTotal = calculate.add(subTotal, total);
-    //item total
-    item['total'] = total;
-  });
-  taxTotal = calculate.multiply(subTotal, taxRate / 100);
-  total = calculate.add(subTotal, taxTotal);
+  const { subTotal, taxTotal, total } = calculateTotals(items, taxRate);
 
   body['subTotal'] = subTotal;
   body['taxTotal'] = taxTotal;
@@ -43,7 +30,6 @@ const create = async (req, res) => {
   body['items'] = items;
 
   let paymentStatus = calculate.sub(total, discount) === 0 ? 'paid' : 'unpaid';
-
   body['paymentStatus'] = paymentStatus;
   body['createdBy'] = req.admin._id;
